🚀 | deploy files from new interaction handler
