+++ src/classes/Client.js
@@ -22,11 +22,11 @@ export class SkyndalexClient extends Client {
 
     async init() {
         await Loaders.loadEvents(this, '../../events');
-
         this.shoukaku = new Shoukaku(new Connectors.DiscordJS(this), Nodes);
         this.shoukaku.on('error', (_, error) => console.error(error));
 
         this.commands = await Loaders.loadCommands('../../commands');
+        this.interactions = await Loaders.loadInteractions('../../interactions');
 
         await this.prisma.$connect();
         await this.login(process.env.BOT_TOKEN);
