+++ src/classes/handlers/Loaders.js
@@ -28,4 +28,25 @@ export class Loaders {
             client.on(name, (...events) => event[name](client, ...events));
         }
     }
+    static async loadInteractions(path) {
+        const interactions = new Collection();
+
+        const files = await readdir(new URL(path, import.meta.url));
+
+        for (const file of files) {
+            const interactionTypes = await readdir(new URL(`${path}/${file}`, import.meta.url));
+            for (const interactionFolders of interactionTypes) {
+                const interactionFile = await readdir(new URL(`${path}/${file}/${interactionFolders}`, import.meta.url));
+
+                for (const interaction of interactionFile) {
+                    const interactionData = await import(`${path}/${file}/${interactionFolders}/${interaction}`);
+                    const customId = interaction.split('.')[0];
+                    if (!interactionData) return console.log(`Interaction ${interaction} does not have a run function`);
+
+                    interactions.set(customId, interactionData);
+                }
+            }
+        }
+        return interactions;
+    }
 }
