+++ src/events/interactionCreate.js
@@ -14,107 +14,31 @@ export async function interactionCreate(client, interaction) {
         )} [#${chalk.yellowBright(interaction.channel?.name)}]`
     );
 
+    console.log(interaction.type);
     switch (interaction.type) {
         case InteractionType.ApplicationCommand:
             if (!interaction.isChatInputCommand()) return;
             const command = client.commands.get(interaction.commandName);
+            if (!command) return interaction.reply("Ooops, this command doesn't exist.");
 
             try {
                 await command.run(client, interaction);
             } catch (error) {
                 console.error(error);
-                await interaction.reply('sorry but i have skill issue');
+                await interaction.reply('Oops, there was an error while executing this command.');
             }
             break;
+        case InteractionType.MessageComponent:
+            const interactionData = client.interactions.get(interaction.customId);
+            if (!interactionData) return interaction.reply("Ooops, this interaction doesn't exist.");
+
+            try {
+                await interactionData.run(client, interaction);
+            } catch (error) {
+                console.error(error);
+                await interaction.reply('Oops, there was an error while executing this interaction.');
+            }
+            break;
+
     }
-    //
-    // if (interaction.isCommand() || interaction.isAutocomplete())
-    //   return await onCommandInteraction(client, interaction);
-    //
-    // const embedNotFound = new EmbedBuilder()
-    //   .setTitle("Not found")
-    //   .setDescription(
-    //     "```❌ | This interaction was not found. Probably this function is refactored, or deleted. Please show this error on the support.```",
-    //   )
-    //   .addFields([
-    //     { name: "Custom ID", value: `\`\`\`${interaction.customId}\`\`\`` },
-    //   ])
-    //   .setColor("Red");
-    //
-    // switch (true) {
-    //   case interaction.isButton():
-    //    {
-    //       const button = client?.interactions
-    //         .filter(
-    //           (x) =>
-    //             x.type === "button" && interaction.customId?.includes(x.customId),
-    //         )
-    //         ?.first();
-    //
-    //       console.log("button", button)
-    //       if (!button)
-    //         return (
-    //           console.log(`Button with ${interaction.customId} was not found.`) ||
-    //           interaction.reply({ embeds: [embedNotFound], ephemeral: true })
-    //         );
-    //       button.run(client, interaction);
-    //     }
-    //     break;
-    //   case interaction.isModalSubmit():
-    //     {
-    //       const modal = client.interactions.find(
-    //         (x) => x.type === "modal" && interaction.customId === x.customId,
-    //       );
-    //       if (!modal)
-    //         return (
-    //           console.log(`Modal with ${interaction.customId} was not found.`) ||
-    //           interaction.reply({ embeds: [embedNotFound] })
-    //         );
-    //
-    //       modal.run(client, interaction);
-    //     }
-    //     break;
-    //   case interaction.isStringSelectMenu():
-    //     const stringSelectOption = client.interactions.find(
-    //       (x) =>
-    //         x.type === "string_select_menu_value" &&
-    //         interaction.customId === x.customId,
-    //     );
-    //
-    //     if (!stringSelectOption)
-    //       return console.log(
-    //         `Select option ${interaction.customId} was not found.`,
-    //       );
-    //
-    //     await stringSelectOption.run(client, interaction);
-    //     break;
-    //   case interaction.isChannelSelectMenu():
-    //     const channelSelectOption = client.interactions.find(
-    //       (x) =>
-    //         x.type === "channel_select_menu_value" &&
-    //         interaction.customId === x.customId,
-    //     );
-    //
-    //     if (!channelSelectOption)
-    //       return console.log(
-    //         `Select option ${interaction.customId} was not found.`,
-    //       );
-    //
-    //     await channelSelectOption.run(client, interaction);
-    //     break;
-    //   case interaction.isRoleSelectMenu():
-    //     const roleSelectOption = client.interactions.find(
-    //       (x) =>
-    //         x.type === "role_select_menu_value" &&
-    //         interaction.customId === x.customId,
-    //     );
-    //
-    //     if (!roleSelectOption)
-    //       return console.log(
-    //         `Select option ${interaction.customId} was not found.`,
-    //       );
-    //
-    //     await roleSelectOption.run(client, interaction);
-    //     break;
-    // }
 }
